cli-sandbox: Add --start option
Rather than define the systemd-nspawn command used to start a user's
sandbox environment within the cli-sandbox service template,
encapsulate it within the cli-sandbox script and expose it via a new
option.

This will allow for more advanced tweaks to be made to the command
arguments, as required, than can be done directly in the template
via specifiers.