cli-sandbox: Replace _ with . in machine names
Since systemd 240 underscores cannot be present in machine names [1].
This poses a problem for the sandbox machines since the username
forms part of the machine name, and we permit underscores in
usernames.

To workaround this we transpose underscores with dots when creating
the user sandbox machines. Since we do not permit dots to be present
in usernames there is no risk of a collision between multiple
distinct users.

VRVDR-49502

[1] https://github.com/systemd/systemd/issues/11765#issuecomment-466961498