Use user ID for sandbox machine names
This is a rework of db143d443d5d512b66e344a1220ef9543864c828 which
addressed the issue of a reduced set of allowed characters in machine
names since systemd 240 (affecting Debian 10+ based releases).

The previous fix was to transpose underscores to periods when
creating a machine for a user's sandbox. This was deemed to be a
complete solution since the DANOS configuration does not permit
periods to be present in usernames, so there was no risk of
collision. Furthermore, other than underscore, all characters we
permit in a username (alphanumerics and dashes) can be used in
machine names.

However this fix did not consider users authenticated by remote AAA
services (eg. TACACS+) which do not have as strict requirements on
usernames as we do for local users. For example a TACACS+ username
could contain periods. If named just so we could place distinct local
and TACACS+ users in the same sandbox if they are named the same
except for transposition of a period and underscore eg. foo.bar vs
foo_bar.

Furthermore in releases without user isolation, or when it is
distabled, we have allowed TACACS+ usernames to contain characters
which are not permitted in machine names but were not (and cannot be)
addressed by the previous fix, such as +.

Any future changes to increase the permitted values for local
usernames, eg. to allow @, would also break the previous fix.

Therefore this commit changes the machine name generation to use the
user's ID rather than their name. Since this is numeric there are no
concerns over whether a machine name is valid or not (max length of
which is 64 characters). The root directory of the sandbox also
changes to match the new machine name but the parent directory
remains as the user's name. For example the sandbox root directory
for user foo with ID 1000 was previously /run/cli-sandbox/foo/cli-foo
but is now /run/cli-sandbox/foo/cli-1000.

This path allows us to verify ownership of the container by building
the expected path with the name and ID of the user who is currently
logging in and checking it matches the RootDirectory of the sandbox
machine (just started or previously running). If the paths do not
match then the machine was created for another user and the login
fails.

This verification is necessary due to the re-use of IDs of deleted
users. There is a very slim possibility of a user being placed in a
sandbox which was created for another user if the original user is
deleted while their sandbox was up, a new user is created and re-uses
the ID of the previous user, and the new user then attempts to log in
while the original sandbox is still up.

We could use other schemes for generating the machine name to prevent
the above scenario, such as a hash of user attributes or by
maintaining a mapping of some kind. However these would add complexity
and reduce readability (eg. of logs) to address a scenario which is very
unlikely to occur, and transient if it does.

VRVDR-49808