cli-sandbox: clean sandboxes of deleted users
If a user is deleted while they have a sandbox running then cleanup
of the sandbox will fail due to the getpwnam() call which fails. This
should not be an issue for local users since the sandbox is stopped
when deleting a local user from the configuration. For users created
via remote AAA methods (eg. TACACS+) it can be the case that users
get deleted before the sandbox is stopped.

Therefore when the sandbox is created we create an environment file
which contains the UID of the user we are creating the sandbox for in
the CLI_SANDBOX_UID variable. If cli-sandbox cannot find the user via
getpwnam() it falls back to obtaining the UID from the environment.
The various sandbox paths can then be built and used to cleanup the
sandbox.

VRVDR-50075