cli-sandbox: Add environment argument to sb_run()
Allows a caller to provide a custom environment for the command to be
executed.

There is no behaviour change if a caller does not provide an env kwarg
to sb_run(), as this simply results in an env kwarg of None to be
passed to subprocess.run(), which is its default value.