cli-sandbox: generate /etc/passwd and /etc/group
Rather than sharing the host /etc/passwd and /etc/group files in
each user sandbox, generate per-sandbox files which contain only
the details relevant for the sandboxed user (plus any base system
users and groups).

This prevents sandboxed users from discovering the identity of
other users on the system through non-modelled methods.

The host "shadow" files, /etc/shadow and /etc/gshadow, are no longer
mounted in sandboxes. In fact the default files generated for the
sandbox chroot are also removed so there are no shadow databases in
sandboxes - they are not required.

The generated passwd and group files are re-generated on a request to
update a given sandbox. Since this occurs on every login, before the
login process enters the sandbox namespaces, this ensures that the
user's shell in the new session is in the appropriate set of groups
(should these have changed since the sandbox was first created).

VRVDR-52842