pam_sandbox.c: reload sandbox on new session
Update pam_sandbox to reload a running sandbox when a new session is
opened. This allows the sandbox to be updated each time a user logs
in.

cli-sandbox@.service is updated to call `cli-sandbox -u` on a request
to reload an instance of the service, and pam_sandbox now performs a
"reload or restart" on this service. If the service is already running
it will be reloaded, otherwise it is started.

To ensure the user's session is not opened until their sandbox has
been updated, enter_sandbox() monitors the modification time of the
sandbox's "ready" file to determine when the sandbox update has
been completed.

If the "ready" file does not exist (eg. because the sandbox has not
yet been started, or is currently being started) the behaviour is
essentially unchanged.

VRVDR-52842