cli-sandbox: Add sandbox update/reload support
This provides the capability to change attributes of a running
sandbox. For example to update files in the sandbox root.

A sandbox update is triggered by calling cli-sandbox with the new
-u/--update argument.

To allow the completion of a sandbox update to be detected in a
simple manner, the modification timestamp of the affected sandbox's
"ready" file (in the root mount namespace) is updated. In fact, as
of this commit, this is all that "updating" a sandbox involves.

VRVDR-52842