cli_sandbox_init: session based keepalive
In a scenario where lots of shortlived sessions are being opened and
sandboxed, cli_sandbox_init may exit (causing the sandbox to stop)
because it believes the sandbox is idle. This causes un-necessary
delay due to stopping/starting of the sandbox.

The sandbox appears idle because the processes of these sessions may
only be alive in between the periodic "number of processes" checks
performed by cli_sandbox_init.

To address this pam_sandbox now updates the modification time of
/run/ready in the sandbox each time a session is placed in the
sandbox. cli_sandbox_init is then adjusted to stay running if a
session was recently (up to 10 seconds prior) placed in the sandbox.

VRVDR-54490