scripts: only bind existing home directory
do not try to bind mount the user's home directory if it doesn't exist
in the system. This can happen with tacplus logins where the home
directory is created at the time of first login, and it may fail if for
some reason the root file system is read-only.

This change allows the isolated users to login without a home directory.

VRVDR-55531